Refactoring started
Some routers have bveen created, but there's still work to do
