+++ app.js
@@ -50,6 +50,13 @@ moment.locale("es");
 var instPool = new pool("localhost", "root", "", "ucm_riu");
 var instDao = new dao(instPool.get_pool());
 
+// Routers import
+const inst_router = require("./routes/r_instalaciones.js");
+const admin_router = require("./routes/r_admin.js");
+const busqueda_router = require("./routes/r_busqueda.js");
+const system_router = require("./routes/r_system.js");
+const correo_router = require("./routes/r_correo.js");
+
 var app = express();
 
 // View engine setup.
@@ -145,703 +152,21 @@ app.get("/no_logged", (request, response, next) => {
   response.status(200).render("must_be_login.ejs");
 });
 
-// Installation management page enrouting.
-app.get("/gestion_instalacion", (request, response) => {
-  if (request.session.isLogged) {
-    if (request.session.is_admin) {
-      var busqueda = "";
-      if (request.query["search"] !== undefined) {
-        busqueda = request.query["search"];
-      }
-      instDao.buscarInstalacion(busqueda, (err, res) => {
-        if (err) {
-          response.status(400).end();
-        } else {
-          res.map((element) => {
-            var [horas, mins, segs] = element.horario_apertura.split(":");
-            element.horario_apertura = horas + ":" + mins;
+// Installation management enrouting.
+app.use("/instalaciones", inst_router);
 
-            var [horas, mins, segs] = element.horario_cierre.split(":");
-            element.horario_cierre = horas + ":" + mins;
-          });
-          response
-            .status(200)
-            .render("gestion_instalaciones.ejs", { dato: res });
-        }
-      });
-    } else {
-      response.status(403).render("no_tienes_permiso.ejs");
-    }
-  } else {
-    response.status(403).render("must_be_login.ejs");
-  }
-});
+// Admin users management enrouting.
+app.use("/admin", admin_router);
 
-// enrouting request to add new instalacion 
-app.post(
-  "/add_instalacion",
-  multerFactory.single("instalacion_imagen"),
-  body("instalacion_nombre").escape(),
-  (request, response) => {
-    if (request.session.isLogged) {
-      if (request.session.is_admin) {
-        var body = request.body;
-        var datos = [
-          body["instalacion_nombre"],
-          body["horario_apertura"],
-          body["horario_cierre"],
-          body["tipo_reserva"],
-          body["aforo"],
-          request.file ? request.file.buffer : null,
-          request.file["mimetype"],
-        ];
-        instDao.insertar_instalacion(datos, (err) => {
-          if (err) {
-            response.status(400);
-          } else {
-            response.status(201).json({
-              msg: "Instalación añadida con éxito, para que haga efecto, refresca la página",
-            });
-          }
-          response.end();
-        });
-      } else {
-        response.status(403).render("no_tienes_permiso.ejs");
-      }
-    } else {
-      response.status(403).render("must_be_login.ejs");
-    }
-  }
-);
+// Admin's advanced search management enrouting.
+app.use("/busqueda", busqueda_router);
 
-// enrouting request to modify an existing installation
-app.put(
-  "/modificar_instalacion/:imagen",
-  multerFactory.single("instalacion_imagen"),
-  body("m_instalacion_nombre").escape(),
-  (request, response) => {
-    if (request.session.isLogged) {
-      if (request.session.is_admin) {
-        var imagen = request.params.imagen == "true" ? true : false;
-        var body = request.body;
-        var id = body["instalacion_id"];
-        var dato = [
-          body["m_instalacion_nombre"],
-          body["m_horario_apertura"],
-          body["m_horario_cierre"],
-          body["m_tipo_reserva"],
-          body["m_aforo"],
-        ];
-        if (imagen) {
-          dato.push(request.file.buffer);
-          dato.push(request.file["mimetype"]);
-        }
-        instDao.modificarInstalacion(id, imagen, dato, (err) => {
-          if (err) {
-            response.status(400);
-          } else {
-            response.status(201).json({
-              msg: "Instalacion modificado con exito, para que se haga efecto, reflesca la pagina",
-            });
-          }
-          response.end();
-        });
-      } else {
-        response.status(403).render("no_tienes_permiso.ejs");
-      }
-    } else {
-      response.status(403).render("must_be_login.ejs");
-    }
-  }
-);
+// Admin's advanced search management enrouting.
+app.use("/system", system_router);
 
-// enrouting request to delete an existing installation
-app.delete("/delete_instalacion/:id", (request, response) => {
-  if (request.session.isLogged) {
-    if (request.session.is_admin) {
-      var id = request.params.id;
-      id = querystring.escape(id);
-      instDao.eliminarInstalacion(id, (err) => {
-        if (err) {
-          response.status(404).json({ msg: "Instalacion no existente" });
-        } else {
-          response.status(201).json({ msg: "Instalacion eliminado con exito" });
-        }
-        response.end();
-      });
-    } else {
-      response.status(403).render("no_tienes_permiso.ejs");
-    }
-  } else {
-    response.status(403).render("must_be_login.ejs");
-  }
-});
+// Inbox and messages management enrouting.
+app.use("/correo", correo_router);
 
-// validate user page enrouting 
-app.get("/validar_registro", (request, response, next) => {
-  if (request.session.isLogged) {
-    if (request.session.is_admin) {
-      instDao.obtenerUsuariosNoValidatos((err, res) => {
-        if (err) {
-          response.status(400).end();
-        } else {
-          response.status(200).render("validarRegistro.ejs", { datos: res });
-        }
-      });
-    } else {
-      response.status(403).render("no_tienes_permiso.ejs");
-    }
-  } else {
-    response.status(403).render("must_be_login.ejs");
-  }
-});
-
-// enrouting request to validate user with the id giving as params.
-app.patch("/validar_registro/:id", (request, response, next) => {
-  if (request.session.isLogged) {
-    if (request.session.is_admin) {
-      var id = request.params.id;
-      id = querystring.escape(id);
-      instDao.validarUsuario(id, (err, res) => {
-        if (err) {
-          response
-            .status(400)
-            .end("Ha ocurrido un error en el acceso interno de la BD.");
-        } else {
-          var datos = [
-            ID_ADMIN,
-            id,
-            `¡Bienvenido a ${request.session.sysConfig.nombre}!`,
-            "Esperamos que disfrutes de nuestras instalaciones. ¡Diviértete con responsabilidad!",
-            false,
-            new Date(),
-          ];
-          instDao.enviarMensaje(datos, (err, res) => {
-            if (err) {
-              response.status(400).end();
-            } else {
-              response.status(200).json({ msg: "Usuario validado con exito" });
-            }
-          });
-        }
-      });
-    } else {
-      response.status(403).render("no_tienes_permiso.ejs");
-    }
-  } else {
-    response.status(403).render("must_be_login.ejs");
-  }
-});
-
-// enrouting request to convert a user with the id giving as params to admin.
-app.patch("/hacer_admin/:id", (request, response, next) => {
-  if (request.session.isLogged) {
-    if (request.session.is_admin) {
-      var id = request.params.id;
-      id = querystring.escape(id);
-      instDao.updateToAdminUser(id, (err, res) => {
-        if (err) {
-          response
-            .status(400)
-            .end("Ha ocurrido un error en el acceso interno de la BD.");
-        } else {
-          var datos = [
-            ID_ADMIN,
-            id,
-            `¡Enhorabuena Te has convertido en admin!`,
-            "Recuerda que: ¡Un gran poder conlleva una gran responsabilidad!",
-            false,
-            new Date(),
-          ];
-          instDao.enviarMensaje(datos, (err, res) => {
-            if (err) {
-              response.status(402).end();
-            } else {
-              response
-                .status(200)
-                .json({ msg: "Usuario convertido en admin." });
-            }
-          });
-        }
-      });
-    } else {
-      response.status(403).render("no_tienes_permiso.ejs");
-    }
-  } else {
-    response.status(403).render("must_be_login.ejs");
-  }
-});
-
-// enrouting request to delete user with the id giving as params.
-app.delete("/eliminar_registro/:id", (request, response, next) => {
-  if (request.session.isLogged) {
-    if (request.session.is_admin) {
-      var id = request.params.id;
-      id = querystring.escape(id);
-      instDao.eliminarUsuario(id, (err, res) => {
-        if (err) {
-          response.status(400).end();
-        } else {
-          response.status(200).json({ msg: "Usuario eliminado con exito" });
-        }
-      });
-    } else {
-      response.status(403).render("no_tienes_permiso.ejs");
-    }
-  } else {
-    response.status(403).render("must_be_login.ejs");
-  }
-});
-
-// User's inbox (all messages received, ordered by date from most recent to oldest).
-app.get("/inbox", (request, response, next) => {
-  if (!request.session.isLogged) {
-    response.status(400).end();
-  } else {
-    instDao.buscarUsuario(request.session.mail, (err, res) => {
-      if (err) {
-        response.status(400).end();
-      } else {
-        var id_receptor = res[0].id;
-        instDao.obtenerMensajes(id_receptor, (err, res) => {
-          if (err) {
-            response.status(400).end();
-          } else {
-            res.forEach((mensaje) => {
-              mensaje.fecha_envio = moment(mensaje.fecha_envio).fromNow();
-            });
-            response.status(200).render("inbox.ejs", { datos: res });
-          }
-        });
-      }
-    });
-  }
-});
-
-// User's messages filtered by a text given by the user.
-app.get("/get_filtered_mail", (request, response, next) => {
-  if (!request.session.isLogged) {
-    response.status(400).end();
-  } else {
-    instDao.buscarUsuario(request.session.mail, (err, res) => {
-      if (err) {
-        response.status(400).end();
-      } else {
-        var id_receptor = res[0].id;
-
-        if (String(request.query["filter_by"]).trimStart().trimEnd() === "") {
-          instDao.obtenerMensajes(id_receptor, (err, res) => {
-            if (err) {
-              response.status(400).end();
-            } else {
-              res.forEach((mensaje) => {
-                mensaje.fecha_envio = moment(mensaje.fecha_envio).fromNow();
-              });
-              response.status(200).render("inbox.ejs", { datos: res });
-            }
-          });
-        } else {
-          instDao.obtenerMensajesFiltrados(
-            id_receptor,
-            request.query["filter_by"],
-            (err, res) => {
-              if (err) {
-                response.status(400).end();
-              } else {
-                res.forEach((mensaje) => {
-                  mensaje.fecha_envio = moment(mensaje.fecha_envio).fromNow();
-                });
-                response.status(200).render("inbox.ejs", { datos: res });
-              }
-            }
-          );
-        }
-      }
-    });
-  }
-});
-
-// system configuration page enrouting.
-app.get("/config_system", (request, response) => {
-  if (request.session.isLogged) {
-    if (request.session.is_admin) {
-      response.status(200).render("config_system.ejs");
-    } else {
-      response.status(403).render("no_tienes_permiso.ejs");
-    }
-  } else {
-    response.status(403).render("must_be_login.ejs");
-  }
-});
-
-// enrouting request to get list of users with the faculty id giving as params.
-app.get("/facultad_usuarios/:id", (request, response) => {
-  if (request.session.isLogged) {
-    if (request.session.is_admin) {
-      var id = request.params.id;
-      id = querystring.escape(id);
-      instDao.obtenerFacultadesListaUsuario(id, (err, result) => {
-        if (err) {
-          response.status(404).end();
-        } else {
-          response.status(200).json({ usuarios: result });
-        }
-      });
-    } else {
-      response.status(403).render("no_tienes_permiso.ejs");
-    }
-  } else {
-    response.status(403).render("must_be_login.ejs");
-  }
-});
-
-// enrouting request to get a list of user history with user id as parameters.
-app.get("/historial_usuario/:id", (request, response) => {
-  if (request.session.isLogged) {
-    if (request.session.is_admin) {
-      var id = request.params.id;
-      id = querystring.escape(id);
-      instDao.obtenerHistorialUsuario(id, (err, result) => {
-        if (err) {
-          response.status(404).end();
-        } else {
-          response.status(200).json({ historial: result });
-        }
-      });
-    } else {
-      response.status(403).render("no_tienes_permiso.ejs");
-    }
-  } else {
-    response.status(403).render("must_be_login.ejs");
-  }
-});
-
-// enrouting request to get a list of reservation history by providing a facility id as parameters.
-app.get("/historial_instalacion/:id", (request, response) => {
-  if (request.session.isLogged) {
-    if (request.session.is_admin) {
-      var id = request.params.id;
-      id = querystring.escape(id);
-      instDao.obtenerHistorialInstalacion(id, (err, result) => {
-        if (err) {
-          response.status(404).end();
-        } else {
-          response.status(200).json({ historial: result });
-        }
-      });
-    } else {
-      response.status(403).render("no_tienes_permiso.ejs");
-    }
-  } else {
-    response.status(403).render("must_be_login.ejs");
-  }
-});
-
-// enrouting request to get user statistics providing user id as parameters.
-app.get("/estadistica_usuario/:id", (request, response) => {
-  if (request.session.isLogged) {
-    if (request.session.is_admin) {
-      var id = request.params.id;
-      id = querystring.escape(id);
-      instDao.obtenerEstadisticaUsuario(id, (err, result) => {
-        if (err) {
-          response.status(404).end();
-        } else {
-          response.status(200).json({ estadistica: result });
-        }
-      });
-    } else {
-      response.status(403).render("no_tienes_permiso.ejs");
-    }
-  } else {
-    response.status(403).render("must_be_login.ejs");
-  }
-});
-
-// enrouting request to get faculty statistics providing faculty id as parameters.
-app.get("/estadistica_facultad/:id", (request, response) => {
-  if (request.session.isLogged) {
-    if (request.session.is_admin) {
-      var id = request.params.id;
-      id = querystring.escape(id);
-      instDao.obtenerEstadisticaFacultad(id, (err, result) => {
-        if (err) {
-          response.status(404).end();
-        } else {
-          response.status(200).json({ estadistica: result });
-        }
-      });
-    } else {
-      response.status(403).render("no_tienes_permiso.ejs");
-    }
-  } else {
-    response.status(403).render("must_be_login.ejs");
-  }
-});
-
-// advanced search page enrouting.
-app.get("/busquedaAvanzada", (request, response, next) => {
-  if (request.session.isLogged) {
-    if (request.session.is_admin) {
-      // get search querys
-      var tipo_busqueda = request.query["tipo_busqueda"];
-      var filtrar_por = request.query["filtrar_por"];
-      var search = request.query["search"];
-      var date_init = request.query["date_init"];
-      var date_end = request.query["date_end"];
-      if (search !== undefined) {
-        search = querystring.escape(search);
-        search = search.replaceAll("%40", "@");
-        search = search.replaceAll("%20", " ");
-      }
-      //callback for each search type.
-      var usuario_callback = (err, res_usuario) => {
-        if (err) {
-          next();
-        } else {
-          response
-            .status(200)
-            .render("busquedaAvanzada.ejs", { usuarios: res_usuario });
-        }
-      };
-      var reservas_callback = (err, res_reservas) => {
-        if (err) {
-          next();
-        } else {
-          response
-            .status(200)
-            .render("busquedaAvanzada.ejs", { reservas: res_reservas });
-        }
-      };
-      var facultad_callback = (err, res_facultad) => {
-        if (err) {
-          next();
-        } else {
-          response
-            .status(200)
-            .render("busquedaAvanzada.ejs", { facultades: res_facultad });
-        }
-      };
-      var instalacion_callback = (err, res_instalacion) => {
-        if (err) {
-          next();
-        } else {
-          response
-            .status(200)
-            .render("busquedaAvanzada.ejs", { instalaciones: res_instalacion });
-        }
-      };
-      //case of search
-      switch (tipo_busqueda) {
-        case "Usuario":
-          switch (filtrar_por) {
-            case "Nombre":
-              instDao.obtenerListaUsuarioPorNombre(search, usuario_callback);
-              break;
-            case "Apellido":
-              instDao.obtenerListaUsuarioPorApellido(search, usuario_callback);
-              break;
-            case "Correo":
-              instDao.obtenerListaUsuarioPorCorreo(search, usuario_callback);
-              break;
-            case "Facultad":
-              instDao.obtenerListaUsuarioPorFaculdad(search, usuario_callback);
-              break;
-          }
-          break;
-        case "Reserva":
-          switch (filtrar_por) {
-            case "Nombre Usuario":
-              instDao.obtenerReservaPorNombreUsuario(search, reservas_callback);
-              break;
-            case "Apellido Usuario":
-              instDao.obtenerReservaPorApellidoUsuario(
-                search,
-                reservas_callback
-              );
-              break;
-            case "Nombre Facultad":
-              instDao.obtenerReservaPorNombreFacultad(
-                search,
-                reservas_callback
-              );
-              break;
-            case "Nombre Instalacion":
-              instDao.obtenerReservaPorNombreInstalacion(
-                search,
-                reservas_callback
-              );
-              break;
-            case "Fecha Inicio - Fecha Fin":
-              instDao.obtenerReservaPorRangoTemporal(
-                [date_init, date_end],
-                reservas_callback
-              );
-              break;
-          }
-          break;
-        case "Facultad":
-          switch (filtrar_por) {
-            case "Nombre Facultad":
-              instDao.obtenerFacultadesPorNombre(search, facultad_callback);
-              break;
-            case "Nombre Usuario":
-              instDao.obtenerFacultadesPorUsuarioNombre(
-                search,
-                facultad_callback
-              );
-              break;
-            case "Apellido Usuario":
-              instDao.obtenerFacultadesPorUsuarioApellido(
-                search,
-                facultad_callback
-              );
-              break;
-          }
-          break;
-        case "Instalacion":
-          instDao.obtenerInstalacionPorNombre(search, instalacion_callback);
-          break;
-        default:
-          response.status(200).render("busquedaAvanzada.ejs");
-      }
-    } else {
-      response.status(403).render("no_tienes_permiso.ejs");
-    }
-  } else {
-    response.status(403).render("must_be_login.ejs");
-  }
-});
-
-// enrouting request to update system configuration.
-app.put(
-  "/update_system/:imagen",
-  multerFactory.single("org_img"),
-  body("org_name").escape(),
-  body("org_dir").escape(),
-  body("org_ig").escape(),
-  body("org_mail").escape(),
-  (request, response, next) => {
-    if (request.session.isLogged) {
-      if (request.session.is_admin) {
-        var imagen = request.params.imagen == "true" ? true : false;
-        var body = request.body;
-        var dato = [
-          body["org_name"],
-          body["org_dir"],
-          body["org_ig"],
-          body["org_mail"],
-        ];
-        if (imagen) {
-          dato.push(request.file.buffer);
-          dato.push(request.file["mimetype"]);
-        }
-        instDao.update_config(imagen, dato, (err) => {
-          if (err) {
-            response.status(400);
-          } else {
-            request.session.sysConfig["nombre"] = dato[0];
-            request.session.sysConfig["direccion"] = dato[1];
-            request.session.sysConfig["url_instagram"] = dato[2].replaceAll(
-              "&#x2F;",
-              "/"
-            );
-            request.session.sysConfig["correo"] = dato[3];
-            if (imagen) {
-              request.session.sysConfig["icono"] = dato[4];
-              request.session.sysConfig["icono_type"] = dato[5];
-            }
-            response
-              .status(201)
-              .json({ msg: "Se ha actualizado correctamente" });
-          }
-          response.end();
-        });
-      } else {
-        response.status(403).render("no_tienes_permiso.ejs");
-      }
-    } else {
-      response.status(403).render("must_be_login.ejs");
-    }
-  }
-);
-
-// Route to mark a message as read (set the boolean to "True").
-app.patch("/marcar_leido/:id", (request, response, next) => {
-  if (!request.session.isLogged) {
-    response.status(400).end();
-  } else {
-    instDao.marcarComoLeido(request.params.id, (err, res) => {
-      if (err) {
-        response
-          .status(400)
-          .end("Ha ocurrido un error en el acceso interno de la BD.");
-      } else {
-        response.status(200).end();
-      }
-    });
-  }
-});
-
-// Route to send a message from a user to another one. It also ensures implied users belong to the same faculty, or at least one of them is admin.
-app.post(
-  "/write_mail",
-  body("mail").escape(),
-  body("receptor").escape(),
-  body("asunto").escape(),
-  (request, response, next) => {
-    if (!request.session.isLogged) {
-      response.status(400).end();
-    } else {
-      instDao.buscarUsuario(request.session.mail, async (err, res) => {
-        if (err) {
-          response.status(400).end();
-        } else {
-          var idEmisor = res[0].id;
-          var emisor = res[0];
-          instDao.buscarUsuario(request.body.receptor, (err, res) => {
-            if (err) {
-              response.status(400).end();
-            } else {
-              // Remitent does not exist.
-              if (res.length == 0) {
-                response.status(400).end("No receiver found");
-              } else {
-                if (
-                  res[0].facultad != emisor.facultad &&
-                  !emisor.es_admin &&
-                  !res[0].es_admin
-                ) 
-                // Different faculty.
-                {
-                  response.status(400).end("Faculty does not match");
-                } else {
-                  var idReceptor = res[0].id;
-
-                  datos = [
-                    idEmisor,
-                    idReceptor,
-                    request.body.asunto,
-                    request.body.mensaje,
-                    false,
-                    new Date(),
-                  ];
-                  instDao.enviarMensaje(datos, (err, res) => {
-                    if (err) {
-                      response.status(400).end();
-                    } else {
-                      response.status(201).end();
-                    }
-                  });
-                }
-              }
-            }
-          });
-        }
-      });
-    }
-  }
-);
 
 // Route to book an installation. If it overlaps with an already booking, throws an erorr to the user indicating that, so that it can decide whether to be
 // added to the wait list or not..
