+++ views/busquedaAvanzada.ejs
@@ -17,7 +17,7 @@
     <main>
         <div class="container-fluid" id="form_container">
             <!-- search form -->
-            <form class="row justify-content-center col-12" method="get" action="/busquedaAvanzada"
+            <form class="row justify-content-center col-12" method="get" action="/busqueda/busquedaAvanzada"
                 id="busqueda_avanzada">
                 <!-- types of search -->
                 <div class="form-group col-md-3 col-sm-5 col-10 d-inline-block m-md-1 m-2">
