+++ views/config_system.ejs
@@ -23,7 +23,7 @@
                 </div>
                 <!-- form to update system configuration like name, instagram, mail etc... -->
                 <div class="col-md-9 col-sm-10 col-12 border border-3 border-success bg-success-subtle rounded-5">
-                    <form id="config_form" method="put" action="/update_system" class="my-3">
+                    <form id="config_form" method="put" action="/system/update_system" class="my-3">
                         <div class="container" id="config_form_body">
                             <!-- Nombre Organizacion -->
                             <div class="form-group row justify-content-center mb-1">
