+++ views/inbox.ejs
@@ -87,7 +87,7 @@
 
                             <!-- When the user searchs something (user name, surname, email content...), a request is made to the server as a form. -->
                             <div class="d-inline-block">
-                                <form class="m-2 d-inline-block" action="/get_filtered_mail"
+                                <form class="m-2 d-inline-block" action="/correo/get_filtered_mail"
                                       method="GET" id="mail_search_bar">
                                     <input id="filter_by" class="search_text rounded-3"
                                         type="text"
