+++ views/pieces/header.ejs
@@ -76,7 +76,7 @@
                         </div>
 
                         <div class="d-flex flex-column mx-3 px-1 justify-content-center ">
-                            <a id="a_inbox" href="/inbox">Bandeja de entrada</a>
+                            <a id="a_inbox" href="/correo/inbox">Bandeja de entrada</a>
                         </div>
                     </div>
                     <% } else { %>
